refactor(auth): :memo:  move authentication decorators to common module
- Consolidate authentication-related decorators (Public, Authenticated, CurrentUser) into the common module to improve code organization and reduce circular dependencies.
- Update all imports across the application to use the new location.
- Add comprehensive documentation in AUTHENTICATION.md.
