+++ src/common/decorators/current-user.decorator.ts
@@ -1,6 +1,9 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
 import { GqlExecutionContext } from '@nestjs/graphql';
-import { JwtPayload } from '../jwt/jwt-handler';
+import { JwtPayload } from '../../domains/authentication/infrastructure/jwt/jwt-handler';
+
+// Re-export JwtPayload for convenience
+export { JwtPayload };
 
 interface RequestWithUser {
   user: JwtPayload;
