+++ src/domains/address/presentation/graphql/address.resolver.ts
@@ -1,6 +1,5 @@
 import { Resolver, Mutation, Args, ID, Query } from '@nestjs/graphql';
-import { CurrentUser } from '@authentication/infrastructure/decorators';
-import { JwtPayload } from '@authentication/infrastructure/jwt';
+import { CurrentUser, JwtPayload } from '@common/decorators';
 import { CommandBus, QueryBus } from '@nestjs/cqrs';
 import {
   AddressType,
