--- src/domains/authentication/infrastructure/decorators/index.ts
@@ -1,3 +1,0 @@
-export { IsPublicKey } from './public.decorator';
-export { IsAuthenticatedKey } from './authenticated.decorator';
-export { CurrentUser } from './current-user.decorator';
